+++ ors-api/src/main/java/org/heigit/ors/api/requests/isochrones/IsochronesRequest.java
@@ -191,11 +191,11 @@ public class IsochronesRequest extends APIRequest {
     public IsochronesRequest() {
     }
 
-    static String[] convertAttributes(IsochronesRequestEnums.Attributes[] attributes) {
+    public static String[] convertAttributes(IsochronesRequestEnums.Attributes[] attributes) {
         return convertAPIEnumListToStrings(attributes);
     }
 
-    protected static int convertToIsochronesProfileType(APIEnums.Profile profile) throws ParameterValueException {
+    public static int convertToIsochronesProfileType(APIEnums.Profile profile) throws ParameterValueException {
         try {
             int profileFromString = RoutingProfileType.getFromString(profile.toString());
             if (profileFromString == 0) {
@@ -370,349 +370,357 @@ public class IsochronesRequest extends APIRequest {
     public boolean hasTime() {
         return hasTime;
     }
-
-    public void generateIsochronesFromRequest(EndpointsProperties endpointsProperties) throws Exception {
-        this.isochroneRequest = this.convertIsochroneRequest(endpointsProperties);
-        // request object is built, now check if ors config allows all settings
-        List<TravellerInfo> travellers = this.isochroneRequest.getTravellers();
-
-        // TODO REFACTORING where should we put the validation code?
-        validateAgainstConfig(this.isochroneRequest, travellers);
-
-        if (!travellers.isEmpty()) {
-            isoMaps = new IsochroneMapCollection();
-
-            for (int i = 0; i < travellers.size(); ++i) {
-                IsochroneSearchParameters searchParams = this.isochroneRequest.getSearchParameters(i);
-                IsochroneMap isochroneMap = RoutingProfileManager.getInstance().buildIsochrone(searchParams);
-                isoMaps.add(isochroneMap);
-            }
-
-        }
-    }
-
-    Float convertSmoothing(Double smoothingValue) throws ParameterValueException {
-        float f = (float) smoothingValue.doubleValue();
-
-        if (smoothingValue < 0 || smoothingValue > 100)
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_SMOOTHING, smoothingValue.toString());
-
-        return f;
-    }
-
-    String convertLocationType(IsochronesRequestEnums.LocationType locationType) throws ParameterValueException {
-        IsochronesRequestEnums.LocationType value;
-
-        switch (locationType) {
-            case DESTINATION:
-                value = IsochronesRequestEnums.LocationType.DESTINATION;
-                break;
-            case START:
-                value = IsochronesRequestEnums.LocationType.START;
-                break;
-            default:
-                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_LOCATION_TYPE, locationType.toString());
-        }
-
-        return value.toString();
-    }
-
-    TravelRangeType convertRangeType(IsochronesRequestEnums.RangeType rangeType) throws ParameterValueException {
-        TravelRangeType travelRangeType;
-
-        switch (rangeType) {
-            case DISTANCE:
-                travelRangeType = TravelRangeType.DISTANCE;
-                break;
-            case TIME:
-                travelRangeType = TravelRangeType.TIME;
-                break;
-            default:
-                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_RANGE_TYPE, rangeType.toString());
-        }
-
-        return travelRangeType;
-
-    }
-
-    String convertAreaUnit(APIEnums.Units unitsIn) throws ParameterValueException {
-
-        DistanceUnit convertedAreaUnit;
-        try {
-            convertedAreaUnit = DistanceUnitUtil.getFromString(unitsIn.toString(), DistanceUnit.UNKNOWN);
-            if (convertedAreaUnit == DistanceUnit.UNKNOWN)
-                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_AREA_UNITS, unitsIn.toString());
-
-            return DistanceUnitUtil.toString(convertedAreaUnit);
-
-        } catch (Exception e) {
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_AREA_UNITS, unitsIn.toString());
-        }
-    }
-
-    String convertRangeUnit(APIEnums.Units unitsIn) throws ParameterValueException {
-
-        DistanceUnit units;
-        try {
-            units = DistanceUnitUtil.getFromString(unitsIn.toString(), DistanceUnit.UNKNOWN);
-            if (units == DistanceUnit.UNKNOWN)
-                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_RANGE_UNITS, unitsIn.toString());
-        } catch (Exception e) {
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_RANGE_UNITS, unitsIn.toString());
-        }
-        return DistanceUnitUtil.toString(units);
-
-    }
-
-    Coordinate convertSingleCoordinate(Double[] coordinate) throws ParameterValueException {
-        Coordinate realCoordinate;
-        if (coordinate.length != 2) {
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_LOCATIONS);
-        }
-        try {
-            realCoordinate = new Coordinate(coordinate[0], coordinate[1]);
-        } catch (Exception e) {
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_LOCATIONS);
-        }
-        return realCoordinate;
-    }
-
-    IsochroneRequest convertIsochroneRequest(EndpointsProperties endpointsProperties) throws Exception {
-        IsochroneRequest convertedIsochroneRequest = new IsochroneRequest();
-        EndpointsProperties.EndpointIsochroneProperties isochroneProperties = endpointsProperties.getIsochrone();
-        convertedIsochroneRequest.setMaximumLocations(isochroneProperties.getMaximumLocations());
-        convertedIsochroneRequest.setAllowComputeArea(isochroneProperties.isAllowComputeArea());
-        convertedIsochroneRequest.setMaximumIntervals(isochroneProperties.getMaximumIntervals());
-        convertedIsochroneRequest.setMaximumRangeDistanceDefault(isochroneProperties.getMaximumRangeDistanceDefault());
-        convertedIsochroneRequest.setProfileMaxRangeDistances(isochroneProperties.getProfileMaxRangeDistances());
-        convertedIsochroneRequest.setMaximumRangeDistanceDefaultFastisochrones(isochroneProperties.getFastisochrones().getMaximumRangeDistanceDefault());
-        convertedIsochroneRequest.setProfileMaxRangeDistancesFastisochrones(isochroneProperties.getFastisochrones().getProfileMaxRangeDistances());
-        convertedIsochroneRequest.setMaximumRangeTimeDefault(isochroneProperties.getMaximumRangeTimeDefault());
-        convertedIsochroneRequest.setProfileMaxRangeTimes(isochroneProperties.getProfileMaxRangeTimes());
-        convertedIsochroneRequest.setMaximumRangeTimeDefaultFastisochrones(isochroneProperties.getFastisochrones().getMaximumRangeTimeDefault());
-        convertedIsochroneRequest.setProfileMaxRangeTimesFastisochrones(isochroneProperties.getFastisochrones().getProfileMaxRangeTimes());
-        convertedIsochroneRequest.setStatsProviders(constructStatisticsProvidersConfiguration(isochroneProperties.getStatisticsProviders()));
-
-        for (int i = 0; i < locations.length; i++) {
-            Double[] location = locations[i];
-            TravellerInfo travellerInfo = this.constructTravellerInfo(location);
-            travellerInfo.setId(Integer.toString(i));
-            try {
-                convertedIsochroneRequest.addTraveller(travellerInfo);
-            } catch (Exception ex) {
-                throw new InternalServerException(IsochronesErrorCodes.UNKNOWN, IsochronesRequest.PARAM_INTERVAL);
-            }
-        }
-        if (this.hasId())
-            convertedIsochroneRequest.setId(this.getId());
-        if (this.hasRangeUnits())
-            convertedIsochroneRequest.setUnits(convertRangeUnit(rangeUnit));
-        if (this.hasAreaUnits())
-            convertedIsochroneRequest.setAreaUnits(convertAreaUnit(areaUnit));
-        if (this.hasAttributes())
-            convertedIsochroneRequest.setAttributes(convertAttributes(attributes));
-        if (this.hasSmoothing())
-            convertedIsochroneRequest.setSmoothingFactor(convertSmoothing(smoothing));
-        if (this.hasIntersections())
-            convertedIsochroneRequest.setIncludeIntersections(intersections);
-        if (this.hasOptions())
-            convertedIsochroneRequest.setCalcMethod(convertCalcMethod(CONCAVE_BALLS));
-        else
-            convertedIsochroneRequest.setCalcMethod(convertCalcMethod(FASTISOCHRONE));
-        return convertedIsochroneRequest;
-
-    }
-
-    Map<String, StatisticsProviderConfiguration> constructStatisticsProvidersConfiguration (Map<String, EndpointsProperties.EndpointIsochroneProperties.StatisticsProviderProperties> statsProperties) {
-        Map<String, StatisticsProviderConfiguration> statsProviders = new HashMap<>();
-
-        if (statsProperties != null) {
-            int id = 0;
-            for (EndpointsProperties.EndpointIsochroneProperties.StatisticsProviderProperties providerProperties : statsProperties.values()) {
-                Map<String, String> propertyMapping = providerProperties.getPropertyMapping();
-                Map<String, String> propMapping = new HashMap<>();
-                for (Map.Entry<String, String> propEntry : propertyMapping.entrySet())
-                    propMapping.put(propEntry.getValue(), propEntry.getKey());
-                if (propMapping.size() > 0) {
-                    StatisticsProviderConfiguration provConfig = new StatisticsProviderConfiguration(id++, providerProperties.getProviderName(), providerProperties.getProviderParameters(), propMapping, providerProperties.getAttribution());
-                    for (Map.Entry<String, String> property : propMapping.entrySet())
-                        statsProviders.put(property.getKey().toLowerCase(), provConfig);
-                }
-            }
-        }
-
-        return statsProviders;
-    }
-
-    TravellerInfo constructTravellerInfo(Double[] coordinate) throws Exception {
-        TravellerInfo travellerInfo = new TravellerInfo();
-
-        RouteSearchParameters routeSearchParameters = this.constructRouteSearchParameters();
-        travellerInfo.setRouteSearchParameters(routeSearchParameters);
-        if (this.hasRangeType())
-            travellerInfo.setRangeType(convertRangeType(rangeType));
-        if (this.hasLocationType())
-            travellerInfo.setLocationType(convertLocationType(locationType));
-        travellerInfo.setLocation(convertSingleCoordinate(coordinate));
-        travellerInfo.getRanges();
-        //range + interval
-        if (range == null) {
-            throw new ParameterValueException(IsochronesErrorCodes.MISSING_PARAMETER, IsochronesRequest.PARAM_RANGE);
-        }
-        List<Double> rangeValues = range;
-        Double intervalValue = interval;
-        setRangeAndIntervals(travellerInfo, rangeValues, intervalValue);
-        return travellerInfo;
-    }
-
-    RouteSearchParameters constructRouteSearchParameters() throws Exception {
-        RouteSearchParameters routeSearchParameters = new RouteSearchParameters();
-        int profileType;
-        try {
-            profileType = convertToIsochronesProfileType(this.getProfile());
-        } catch (Exception e) {
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_PROFILE);
-        }
-
-        if (profileType == RoutingProfileType.UNKNOWN)
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_PROFILE);
-        routeSearchParameters.setProfileType(profileType);
-
-        if (this.hasOptions()) {
-            routeSearchParameters = this.processIsochronesRequestOptions(routeSearchParameters);
-        }
-        if (this.hasTime()) {
-            routeSearchParameters.setDeparture(this.getTime());
-            routeSearchParameters.setArrival(this.getTime());
-        }
-        routeSearchParameters.setConsiderTurnRestrictions(false);
-        return routeSearchParameters;
-    }
-
-    RouteSearchParameters processIsochronesRequestOptions(RouteSearchParameters parameters) throws StatusCodeException {
-        RouteRequestOptions options = isochronesOptions;
-        parameters = this.processRequestOptions(options, parameters);
-        if (options.hasProfileParams())
-            parameters.setProfileParams(convertParameters(options, parameters.getProfileType()));
-        return parameters;
-    }
-
-    void validateAgainstConfig(IsochroneRequest isochroneRequest, List<TravellerInfo> travellers) throws StatusCodeException {
-        if (!isochroneRequest.isAllowComputeArea() && isochroneRequest.hasAttribute("area"))
-            throw new StatusCodeException(StatusCode.BAD_REQUEST, IsochronesErrorCodes.FEATURE_NOT_SUPPORTED, "Area computation is not enabled.");
-
-        if (travellers.size() > isochroneRequest.getMaximumLocations())
-            throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM, IsochronesRequest.PARAM_LOCATIONS, Integer.toString(travellers.size()), Integer.toString(isochroneRequest.getMaximumLocations()));
-
-        for (TravellerInfo traveller : travellers) {
-            int maxAllowedRange = getMaximumRange(traveller, isochroneRequest);
-            double maxRange = traveller.getMaximumRange();
-            if (maxRange > maxAllowedRange)
-                throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM, IsochronesRequest.PARAM_RANGE, Double.toString(maxRange), Integer.toString(maxAllowedRange));
-
-            int maxIntervals = isochroneRequest.getMaximumIntervals();
-            if (maxIntervals > 0 && maxIntervals < traveller.getRanges().length) {
-                throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MINIMUM, IsochronesRequest.PARAM_INTERVAL, "Resulting number of " + traveller.getRanges().length + " isochrones exceeds maximum value of " + maxIntervals + ".");
-            }
-        }
-
-    }
-
-    private static int getMaximumRange(TravellerInfo traveller, IsochroneRequest isochroneRequest) {
-        int profileType = traveller.getRouteSearchParameters().getProfileType();
-        TravelRangeType range = traveller.getRangeType();
-        String calcMethod = isochroneRequest.getCalcMethod();
-        Integer res;
-
-        RoutingProfileManager rpm = RoutingProfileManager.getInstance();
-        FastIsochroneFactory fastIsochroneFactory = rpm.getProfiles().getRouteProfile(profileType).getGraphhopper().getFastIsochroneFactory();
-        if (fastIsochroneFactory.isEnabled() && calcMethod.equalsIgnoreCase("fastisochrone"))
-            return getMaximumRangeFastIsochrone(traveller, isochroneRequest);
-
-        if (range == DISTANCE) {
-            res = isochroneRequest.getProfileMaxRangeDistances().get(profileType);
-            if (res == null)
-                res = isochroneRequest.getMaximumRangeDistanceDefault();
-        } else {
-            res = isochroneRequest.getProfileMaxRangeTimes().get(profileType);
-            if (res == null)
-                res = isochroneRequest.getMaximumRangeTimeDefault();
-        }
-
-        return res;
-    }
-
-    private static int getMaximumRangeFastIsochrone(TravellerInfo traveller, IsochroneRequest isochroneRequest) {
-        int profileType = traveller.getRouteSearchParameters().getProfileType();
-        TravelRangeType range = traveller.getRangeType();
-        Integer res;
-
-        if (range == DISTANCE) {
-            res = isochroneRequest.getProfileMaxRangeDistancesFastisochrones().get(profileType);
-            if (res == null)
-                res = isochroneRequest.getMaximumRangeDistanceDefaultFastisochrones();
-        } else {
-            res = isochroneRequest.getProfileMaxRangeTimesFastisochrones().get(profileType);
-            if (res == null)
-                res = isochroneRequest.getMaximumRangeTimeDefaultFastisochrones();
-        }
-
-        return res;
-    }
-
-    void setRangeAndIntervals(TravellerInfo travellerInfo, List<Double> rangeValues, Double intervalValue) throws ParameterValueException, ParameterOutOfRangeException {
-        double rangeValue = -1;
-        if (rangeValues.size() == 1) {
-            try {
-                rangeValue = rangeValues.get(0);
-                travellerInfo.setRanges(new double[]{rangeValue});
-            } catch (NumberFormatException ex) {
-                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, "range");
-            }
-        } else {
-            double[] ranges = new double[rangeValues.size()];
-            double maxRange = Double.MIN_VALUE;
-            for (int i = 0; i < ranges.length; i++) {
-                double dv = rangeValues.get(i);
-                if (dv > maxRange)
-                    maxRange = dv;
-                ranges[i] = dv;
-            }
-            Arrays.sort(ranges);
-            travellerInfo.setRanges(ranges);
-        }
-        // interval, only use if one range is defined
-
-        if (rangeValues.size() == 1 && rangeValue != -1 && intervalValue != null) {
-            if (intervalValue > rangeValue) {
-                throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM, IsochronesRequest.PARAM_INTERVAL, Double.toString(intervalValue), Double.toString(rangeValue));
-            }
-            travellerInfo.setRanges(rangeValue, intervalValue);
-        }
-    }
-
-    String convertCalcMethod(IsochronesRequestEnums.CalculationMethod bareCalcMethod) throws ParameterValueException {
-        try {
-            switch (bareCalcMethod) {
-                case CONCAVE_BALLS:
-                    return "concaveballs";
-                case GRID:
-                    return "grid";
-                case FASTISOCHRONE:
-                    return "fastisochrone";
-                default:
-                    return "none";
-            }
-        } catch (Exception ex) {
-            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, "calc_method");
-        }
-    }
+//
+//    public void generateIsochronesFromRequest(EndpointsProperties endpointsProperties) throws Exception {
+//        this.isochroneRequest = this.convertIsochroneRequest(endpointsProperties);
+//        // request object is built, now check if ors config allows all settings
+//        List<TravellerInfo> travellers = this.isochroneRequest.getTravellers();
+//
+//        // TODO REFACTORING where should we put the validation code?
+//        validateAgainstConfig(this.isochroneRequest, travellers);
+//
+//        if (!travellers.isEmpty()) {
+//            isoMaps = new IsochroneMapCollection();
+//
+//            for (int i = 0; i < travellers.size(); ++i) {
+//                IsochroneSearchParameters searchParams = this.isochroneRequest.getSearchParameters(i);
+//                IsochroneMap isochroneMap = RoutingProfileManager.getInstance().buildIsochrone(searchParams);
+//                isoMaps.add(isochroneMap);
+//            }
+//
+//        }
+//    }
+//
+//    Float convertSmoothing(Double smoothingValue) throws ParameterValueException {
+//        float f = (float) smoothingValue.doubleValue();
+//
+//        if (smoothingValue < 0 || smoothingValue > 100)
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_SMOOTHING, smoothingValue.toString());
+//
+//        return f;
+//    }
+//
+//    String convertLocationType(IsochronesRequestEnums.LocationType locationType) throws ParameterValueException {
+//        IsochronesRequestEnums.LocationType value;
+//
+//        switch (locationType) {
+//            case DESTINATION:
+//                value = IsochronesRequestEnums.LocationType.DESTINATION;
+//                break;
+//            case START:
+//                value = IsochronesRequestEnums.LocationType.START;
+//                break;
+//            default:
+//                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_LOCATION_TYPE, locationType.toString());
+//        }
+//
+//        return value.toString();
+//    }
+//
+//    TravelRangeType convertRangeType(IsochronesRequestEnums.RangeType rangeType) throws ParameterValueException {
+//        TravelRangeType travelRangeType;
+//
+//        switch (rangeType) {
+//            case DISTANCE:
+//                travelRangeType = TravelRangeType.DISTANCE;
+//                break;
+//            case TIME:
+//                travelRangeType = TravelRangeType.TIME;
+//                break;
+//            default:
+//                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_RANGE_TYPE, rangeType.toString());
+//        }
+//
+//        return travelRangeType;
+//
+//    }
+//
+//    String convertAreaUnit(APIEnums.Units unitsIn) throws ParameterValueException {
+//
+//        DistanceUnit convertedAreaUnit;
+//        try {
+//            convertedAreaUnit = DistanceUnitUtil.getFromString(unitsIn.toString(), DistanceUnit.UNKNOWN);
+//            if (convertedAreaUnit == DistanceUnit.UNKNOWN)
+//                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_AREA_UNITS, unitsIn.toString());
+//
+//            return DistanceUnitUtil.toString(convertedAreaUnit);
+//
+//        } catch (Exception e) {
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_AREA_UNITS, unitsIn.toString());
+//        }
+//    }
+//
+//    String convertRangeUnit(APIEnums.Units unitsIn) throws ParameterValueException {
+//
+//        DistanceUnit units;
+//        try {
+//            units = DistanceUnitUtil.getFromString(unitsIn.toString(), DistanceUnit.UNKNOWN);
+//            if (units == DistanceUnit.UNKNOWN)
+//                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_RANGE_UNITS, unitsIn.toString());
+//        } catch (Exception e) {
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_RANGE_UNITS, unitsIn.toString());
+//        }
+//        return DistanceUnitUtil.toString(units);
+//
+//    }
+//
+//    Coordinate convertSingleCoordinate(Double[] coordinate) throws ParameterValueException {
+//        Coordinate realCoordinate;
+//        if (coordinate.length != 2) {
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_LOCATIONS);
+//        }
+//        try {
+//            realCoordinate = new Coordinate(coordinate[0], coordinate[1]);
+//        } catch (Exception e) {
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_LOCATIONS);
+//        }
+//        return realCoordinate;
+//    }
+//
+//    IsochroneRequest convertIsochroneRequest(EndpointsProperties endpointsProperties) throws Exception {
+//        IsochroneRequest convertedIsochroneRequest = new IsochroneRequest();
+//        EndpointsProperties.EndpointIsochroneProperties isochroneProperties = endpointsProperties.getIsochrone();
+//        convertedIsochroneRequest.setMaximumLocations(isochroneProperties.getMaximumLocations());
+//        convertedIsochroneRequest.setAllowComputeArea(isochroneProperties.isAllowComputeArea());
+//        convertedIsochroneRequest.setMaximumIntervals(isochroneProperties.getMaximumIntervals());
+//        convertedIsochroneRequest.setMaximumRangeDistanceDefault(isochroneProperties.getMaximumRangeDistanceDefault());
+//        convertedIsochroneRequest.setProfileMaxRangeDistances(isochroneProperties.getProfileMaxRangeDistances());
+//        convertedIsochroneRequest.setMaximumRangeDistanceDefaultFastisochrones(isochroneProperties.getFastisochrones().getMaximumRangeDistanceDefault());
+//        convertedIsochroneRequest.setProfileMaxRangeDistancesFastisochrones(isochroneProperties.getFastisochrones().getProfileMaxRangeDistances());
+//        convertedIsochroneRequest.setMaximumRangeTimeDefault(isochroneProperties.getMaximumRangeTimeDefault());
+//        convertedIsochroneRequest.setProfileMaxRangeTimes(isochroneProperties.getProfileMaxRangeTimes());
+//        convertedIsochroneRequest.setMaximumRangeTimeDefaultFastisochrones(isochroneProperties.getFastisochrones().getMaximumRangeTimeDefault());
+//        convertedIsochroneRequest.setProfileMaxRangeTimesFastisochrones(isochroneProperties.getFastisochrones().getProfileMaxRangeTimes());
+//        convertedIsochroneRequest.setStatsProviders(constructStatisticsProvidersConfiguration(isochroneProperties.getStatisticsProviders()));
+//
+//        for (int i = 0; i < locations.length; i++) {
+//            Double[] location = locations[i];
+//            TravellerInfo travellerInfo = this.constructTravellerInfo(location);
+//            travellerInfo.setId(Integer.toString(i));
+//            try {
+//                convertedIsochroneRequest.addTraveller(travellerInfo);
+//            } catch (Exception ex) {
+//                throw new InternalServerException(IsochronesErrorCodes.UNKNOWN, IsochronesRequest.PARAM_INTERVAL);
+//            }
+//        }
+//        if (this.hasId())
+//            convertedIsochroneRequest.setId(this.getId());
+//        if (this.hasRangeUnits())
+//            convertedIsochroneRequest.setUnits(convertRangeUnit(rangeUnit));
+//        if (this.hasAreaUnits())
+//            convertedIsochroneRequest.setAreaUnits(convertAreaUnit(areaUnit));
+//        if (this.hasAttributes())
+//            convertedIsochroneRequest.setAttributes(convertAttributes(attributes));
+//        if (this.hasSmoothing())
+//            convertedIsochroneRequest.setSmoothingFactor(convertSmoothing(smoothing));
+//        if (this.hasIntersections())
+//            convertedIsochroneRequest.setIncludeIntersections(intersections);
+//        if (this.hasOptions())
+//            convertedIsochroneRequest.setCalcMethod(convertCalcMethod(CONCAVE_BALLS));
+//        else
+//            convertedIsochroneRequest.setCalcMethod(convertCalcMethod(FASTISOCHRONE));
+//        return convertedIsochroneRequest;
+//
+//    }
+//
+//    Map<String, StatisticsProviderConfiguration> constructStatisticsProvidersConfiguration (Map<String, EndpointsProperties.EndpointIsochroneProperties.StatisticsProviderProperties> statsProperties) {
+//        Map<String, StatisticsProviderConfiguration> statsProviders = new HashMap<>();
+//
+//        if (statsProperties != null) {
+//            int id = 0;
+//            for (EndpointsProperties.EndpointIsochroneProperties.StatisticsProviderProperties providerProperties : statsProperties.values()) {
+//                Map<String, String> propertyMapping = providerProperties.getPropertyMapping();
+//                Map<String, String> propMapping = new HashMap<>();
+//                for (Map.Entry<String, String> propEntry : propertyMapping.entrySet())
+//                    propMapping.put(propEntry.getValue(), propEntry.getKey());
+//                if (propMapping.size() > 0) {
+//                    StatisticsProviderConfiguration provConfig = new StatisticsProviderConfiguration(id++, providerProperties.getProviderName(), providerProperties.getProviderParameters(), propMapping, providerProperties.getAttribution());
+//                    for (Map.Entry<String, String> property : propMapping.entrySet())
+//                        statsProviders.put(property.getKey().toLowerCase(), provConfig);
+//                }
+//            }
+//        }
+//
+//        return statsProviders;
+//    }
+//
+//    TravellerInfo constructTravellerInfo(Double[] coordinate) throws Exception {
+//        TravellerInfo travellerInfo = new TravellerInfo();
+//
+//        RouteSearchParameters routeSearchParameters = this.constructRouteSearchParameters();
+//        travellerInfo.setRouteSearchParameters(routeSearchParameters);
+//        if (this.hasRangeType())
+//            travellerInfo.setRangeType(convertRangeType(rangeType));
+//        if (this.hasLocationType())
+//            travellerInfo.setLocationType(convertLocationType(locationType));
+//        travellerInfo.setLocation(convertSingleCoordinate(coordinate));
+//        travellerInfo.getRanges();
+//        //range + interval
+//        if (range == null) {
+//            throw new ParameterValueException(IsochronesErrorCodes.MISSING_PARAMETER, IsochronesRequest.PARAM_RANGE);
+//        }
+//        List<Double> rangeValues = range;
+//        Double intervalValue = interval;
+//        setRangeAndIntervals(travellerInfo, rangeValues, intervalValue);
+//        return travellerInfo;
+//    }
+//
+//    RouteSearchParameters constructRouteSearchParameters() throws Exception {
+//        RouteSearchParameters routeSearchParameters = new RouteSearchParameters();
+//        int profileType;
+//        try {
+//            profileType = convertToIsochronesProfileType(this.getProfile());
+//        } catch (Exception e) {
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_PROFILE);
+//        }
+//
+//        if (profileType == RoutingProfileType.UNKNOWN)
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, IsochronesRequest.PARAM_PROFILE);
+//        routeSearchParameters.setProfileType(profileType);
+//
+//        if (this.hasOptions()) {
+//            routeSearchParameters = this.processIsochronesRequestOptions(routeSearchParameters);
+//        }
+//        if (this.hasTime()) {
+//            routeSearchParameters.setDeparture(this.getTime());
+//            routeSearchParameters.setArrival(this.getTime());
+//        }
+//        routeSearchParameters.setConsiderTurnRestrictions(false);
+//        return routeSearchParameters;
+//    }
+//
+//    RouteSearchParameters processIsochronesRequestOptions(RouteSearchParameters parameters) throws StatusCodeException {
+//        RouteRequestOptions options = isochronesOptions;
+//        parameters = this.processRequestOptions(options, parameters);
+//        if (options.hasProfileParams())
+//            parameters.setProfileParams(convertParameters(options, parameters.getProfileType()));
+//        return parameters;
+//    }
+//
+//    void validateAgainstConfig(IsochroneRequest isochroneRequest, List<TravellerInfo> travellers) throws StatusCodeException {
+//        if (!isochroneRequest.isAllowComputeArea() && isochroneRequest.hasAttribute("area"))
+//            throw new StatusCodeException(StatusCode.BAD_REQUEST, IsochronesErrorCodes.FEATURE_NOT_SUPPORTED, "Area computation is not enabled.");
+//
+//        if (travellers.size() > isochroneRequest.getMaximumLocations())
+//            throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM, IsochronesRequest.PARAM_LOCATIONS, Integer.toString(travellers.size()), Integer.toString(isochroneRequest.getMaximumLocations()));
+//
+//        for (TravellerInfo traveller : travellers) {
+//            int maxAllowedRange = getMaximumRange(traveller, isochroneRequest);
+//            double maxRange = traveller.getMaximumRange();
+//            if (maxRange > maxAllowedRange)
+//                throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM, IsochronesRequest.PARAM_RANGE, Double.toString(maxRange), Integer.toString(maxAllowedRange));
+//
+//            int maxIntervals = isochroneRequest.getMaximumIntervals();
+//            if (maxIntervals > 0 && maxIntervals < traveller.getRanges().length) {
+//                throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MINIMUM, IsochronesRequest.PARAM_INTERVAL, "Resulting number of " + traveller.getRanges().length + " isochrones exceeds maximum value of " + maxIntervals + ".");
+//            }
+//        }
+//
+//    }
+//
+//    private static int getMaximumRange(TravellerInfo traveller, IsochroneRequest isochroneRequest) {
+//        int profileType = traveller.getRouteSearchParameters().getProfileType();
+//        TravelRangeType range = traveller.getRangeType();
+//        String calcMethod = isochroneRequest.getCalcMethod();
+//        Integer res;
+//
+//        RoutingProfileManager rpm = RoutingProfileManager.getInstance();
+//        FastIsochroneFactory fastIsochroneFactory = rpm.getProfiles().getRouteProfile(profileType).getGraphhopper().getFastIsochroneFactory();
+//        if (fastIsochroneFactory.isEnabled() && calcMethod.equalsIgnoreCase("fastisochrone"))
+//            return getMaximumRangeFastIsochrone(traveller, isochroneRequest);
+//
+//        if (range == DISTANCE) {
+//            res = isochroneRequest.getProfileMaxRangeDistances().get(profileType);
+//            if (res == null)
+//                res = isochroneRequest.getMaximumRangeDistanceDefault();
+//        } else {
+//            res = isochroneRequest.getProfileMaxRangeTimes().get(profileType);
+//            if (res == null)
+//                res = isochroneRequest.getMaximumRangeTimeDefault();
+//        }
+//
+//        return res;
+//    }
+//
+//    private static int getMaximumRangeFastIsochrone(TravellerInfo traveller, IsochroneRequest isochroneRequest) {
+//        int profileType = traveller.getRouteSearchParameters().getProfileType();
+//        TravelRangeType range = traveller.getRangeType();
+//        Integer res;
+//
+//        if (range == DISTANCE) {
+//            res = isochroneRequest.getProfileMaxRangeDistancesFastisochrones().get(profileType);
+//            if (res == null)
+//                res = isochroneRequest.getMaximumRangeDistanceDefaultFastisochrones();
+//        } else {
+//            res = isochroneRequest.getProfileMaxRangeTimesFastisochrones().get(profileType);
+//            if (res == null)
+//                res = isochroneRequest.getMaximumRangeTimeDefaultFastisochrones();
+//        }
+//
+//        return res;
+//    }
+//
+//    void setRangeAndIntervals(TravellerInfo travellerInfo, List<Double> rangeValues, Double intervalValue) throws ParameterValueException, ParameterOutOfRangeException {
+//        double rangeValue = -1;
+//        if (rangeValues.size() == 1) {
+//            try {
+//                rangeValue = rangeValues.get(0);
+//                travellerInfo.setRanges(new double[]{rangeValue});
+//            } catch (NumberFormatException ex) {
+//                throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, "range");
+//            }
+//        } else {
+//            double[] ranges = new double[rangeValues.size()];
+//            double maxRange = Double.MIN_VALUE;
+//            for (int i = 0; i < ranges.length; i++) {
+//                double dv = rangeValues.get(i);
+//                if (dv > maxRange)
+//                    maxRange = dv;
+//                ranges[i] = dv;
+//            }
+//            Arrays.sort(ranges);
+//            travellerInfo.setRanges(ranges);
+//        }
+//        // interval, only use if one range is defined
+//
+//        if (rangeValues.size() == 1 && rangeValue != -1 && intervalValue != null) {
+//            if (intervalValue > rangeValue) {
+//                throw new ParameterOutOfRangeException(IsochronesErrorCodes.PARAMETER_VALUE_EXCEEDS_MAXIMUM, IsochronesRequest.PARAM_INTERVAL, Double.toString(intervalValue), Double.toString(rangeValue));
+//            }
+//            travellerInfo.setRanges(rangeValue, intervalValue);
+//        }
+//    }
+//
+//    String convertCalcMethod(IsochronesRequestEnums.CalculationMethod bareCalcMethod) throws ParameterValueException {
+//        try {
+//            switch (bareCalcMethod) {
+//                case CONCAVE_BALLS:
+//                    return "concaveballs";
+//                case GRID:
+//                    return "grid";
+//                case FASTISOCHRONE:
+//                    return "fastisochrone";
+//                default:
+//                    return "none";
+//            }
+//        } catch (Exception ex) {
+//            throw new ParameterValueException(IsochronesErrorCodes.INVALID_PARAMETER_VALUE, "calc_method");
+//        }
+//    }
 
     public IsochroneMapCollection getIsoMaps() {
         return isoMaps;
     }
 
+    public void setIsoMaps(IsochroneMapCollection isoMaps) {
+        this.isoMaps = isoMaps;
+    }
+
     public IsochroneRequest getIsochroneRequest() {
         return isochroneRequest;
+    }
+
+    public void setIsochroneRequest(IsochroneRequest isochroneRequest) {
+        this.isochroneRequest = isochroneRequest;
     }
 }
